requirements.txt
# Web Framework
streamlit>=1.28.0

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Search APIs
duckduckgo_search>=4.0.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.0  # Token counting for OpenAI
